Implement def!
The Eval struct is instantiated with it's own variable -> AstNode
environment mapping. It's core method, eval(), takes a reference to an
AST parse tree, which it walks and evaluates returning new AstNodes with
intermediate results and lifetimes separate from the given parse tree.

The environment mapping is used by def! to assign variable names to
values. def! is primarily called for its side effects, but must also
return a value, so it has been defined to evaluate to nil.

The lifetimes in the environment mapping were hard to get right. It maps
from a &str with a lifetime equivalent to or greater than the nodes in
the parse tree, since it reuses the variable name strings in the parse
tree as keys. It maps to an AstNode with a lifetime separate from that
of the parse tree (i.e. an intermediate node from eval).